Fixed indentation error in the block.py file

+++ block.py
@@ -22,15 +22,14 @@ class Block:
         else:
             userId = user_id[0][0]
 
-        blockId = dbhelpers.run_insert_statement(
-            "INSERT INTO block (user_id, blocked_user_id) VALUES (?, ?)", [
-                userId, blocked_userId]
-        )
+            blockId = dbhelpers.run_insert_statement(
+                "INSERT INTO block (user_id, blocked_user_id) VALUES (?, ?)", [userId, blocked_userId]
+            )
 
-        if(blockId == None):
-            return Response("Failed to block user", mimetype="text/plain", status=500)
-        else:
-            return Response(status=204)
+            if(blockId == None):
+                return Response("Failed to block user", mimetype="text/plain", status=500)
+            else:
+                return Response(status=204)
 
     # Unblock User
 
@@ -48,12 +47,11 @@ class Block:
         else:
             userId = user_id[0][0]
 
-        rowcount = dbhelpers.run_delete_statement(
-            "DELETE from block b WHERE b.user_id = ? AND b.blocked_user_id = ?", [
-                userId, blocked_userId]
-        )
+            rowcount = dbhelpers.run_delete_statement(
+                "DELETE from block b WHERE b.user_id = ? AND b.blocked_user_id = ?", [userId, blocked_userId]
+            )
 
-        if(rowcount == 1):
-            return Response(status=204)
-        elif(rowcount == None):
-            return Response("Failed to unblock user", mimetype="text/plain", status=500)
+            if(rowcount == 1):
+                return Response(status=204)
+            elif(rowcount == None):
+                return Response("Failed to unblock user", mimetype="text/plain", status=500)
